Refactor server utils into login command
